requirements.txt
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==22.1
sniffio==1.3.1
urllib3==2.4.0
